I am now able to run the GitHub and I added a basic dropdown.
I had issues with dependencies and I had errors every time I tried to run the project. I can now run the project but there are still problems installed packages using npm. I added a working dropdown to the Search Bar but it requires HeadlessUI so I will have to get npm working on the main part to add it to our main project. This dropdown is in a test folder just to get things working. Also fixed the Search Bar formatting.