Added the map to the home page
There is now a map on the homepage that when clicked will replace all of the listings with the map. The map now shows up centered on the screen instead of on the right of the screen. Modified the code so that the map file can access all of the listings.